# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api_rest.errorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  api_rest.fileListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/backend_internal_files.FileListItem'
        type: array
    type: object
...
  api_rest.messageResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
...
  api_rest.quotaUsageResponse:
    properties:
      deduped_bytes:
        type: integer
      original_bytes:
        type: integer
      quota_bytes:
        type: integer
      quota_used_percent:
        type: number
      savings_bytes:
        type: integer
      savings_percent:
        type: number
    type: object
...
  api_rest.searchFileItem:
    properties:
      contentHash:
        type: string
      createdAt:
        type: string
      dedupSavings:
        type: integer
      downloadCount:
        type: integer
      filename:
        type: string
      id:
        type: string
      mime:
        type: string
      physicalSize:
        type: integer
      rank:
        type: number
      refCount:
        type: integer
      size:
        type: integer
      updatedAt:
        type: string
    type: object
  api_rest.searchFilesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/api_rest.searchFileItem'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  backend_internal_files.FileListItem:
    properties:
      contentHash:
        type: string
      createdAt:
        type: string
      dedupSavings:
        type: integer
      deletedAt:
        type: string
      downloadCount:
        type: integer
      filename:
        type: string
      id:
        type: string
      mime:
        type: string
      physicalSize:
        type: integer
      refCount:
        type: integer
      size:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://example.com/support
  description: REST, GraphQL, and WebSocket endpoints for the file service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BalkanID File Service API
  version: "1.0"
paths:
...
  /api/v1/files:
    get:
      consumes:
      - application/json
      description: List files with optional folder filtering and trash support
      parameters:
      - description: Filter by folder ID
        in: query
        name: folderId
        type: string
      - description: List deleted/trash files
        in: query
        name: deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Files list response
          schema:
            $ref: '#/definitions/api_rest.fileListResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List files
      tags:
      - files
...
  /api/v1/search/files:
    get:
      description: Search for files using various filters including query text, MIME
        type, folder location, and uploader
      parameters:
      - description: Search query text for file names and content
        in: query
        name: q
        type: string
      - description: Filter by MIME type (e.g., image/jpeg, text/plain)
        in: query
        name: mime
        type: string
      - description: Filter by parent folder ID
        in: query
        name: folderId
        type: string
      - description: Filter by uploader user ID
        in: query
        name: uploader
        type: string
      - description: Sort order (name, size, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: 'Maximum results to return (1-200, default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Search results with files array and metadata
          schema:
            $ref: '#/definitions/api_rest.searchFilesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      security:
      - BearerAuth: []
      summary: Search files across the system
      tags:
      - search
...
  /api/v1/users/{id}/quota:
    patch:
      consumes:
      - application/json
      description: Update the storage quota limit for a specific user (requires admin
        privileges)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Quota update request
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Quota updated successfully
          schema:
            $ref: '#/definitions/api_rest.messageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      security:
      - BearerAuth: []
      summary: Update user quota (admin only)
      tags:
      - quota
  /api/v1/users/me/usage:
    get:
      description: Retrieve quota usage statistics for the authenticated user including
        storage metrics and savings
      produces:
      - application/json
      responses:
        "200":
          description: Quota usage details
          schema:
            $ref: '#/definitions/api_rest.quotaUsageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api_rest.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user quota usage
      tags:
      - quota
...
securityDefinitions:
  BearerAuth:
    description: 'Provide a Bearer token: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"